requirements.txt
absl_py==1.4.0
aiofiles==23.2.1
dynaconf==3.2.0
efficientnet==1.1.1
fastapi==0.95.2
h5py==3.9.0
httpx==0.24.1
imantics==0.1.12
imgaug==0.4.0
ipython==8.14.0
Jinja2==3.1.2
keras==2.13.1
loguru==0.7.0
lxml==4.9.3
matplotlib==3.7.2
minio==7.1.15
numpy==1.24.3
opencv_python==4.8.0.76
pdf2image==1.16.3
Pillow==10.0.0
pillow_heif==0.13.0
pydantic==1.10.11
pystow==0.5.0
PyYAML==6.0.1
rdkit==2023.3.2
Requests==2.31.0
scipy==1.8.1
selfies==2.1.1
six==1.16.0
scikit-image==0.21.0
starlette==0.27.0
tensorflow==2.13.0
tensorflow_addons==0.21.0
urllib3==1.26.16
uvicorn==0.23.2
python-multipart==0.0.6